java/src/test/java/_20220608_ArrayIntervalStd/SolutionTest.java
package _20220608_ArrayIntervalStd;

import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.junit.runners.Parameterized;

@RunWith(Parameterized.class)
public class SolutionTest {

    private Solution solution;

    @Parameterized.Parameter
    public int[] nums;

    @Parameterized.Parameter(1)
    public int[][] queries;

    @Parameterized.Parameter(2)
    public double[] expected;

    @Parameterized.Parameters
    public static Object[][] data() {
        return new Object[][]{
                {new int[]{1, 3, 1, 2, 1}, new int[][]{{1, 3}, {2, 4}}, new double[]{0.88889, 0.66667}},
        };
    }

    @Before
    public void setup() {
        solution = new Solution();
    }

    @Test
    public void test() {
        double[] result = solution.array_interval_std(nums, queries);
        Assert.assertEquals(expected.length, result.length);
        for (int i = 0; i < expected.length; i++) {
            Assert.assertEquals(expected[i], result[i], 0.00001);
        }
    }

}